wip: styles in searchbar autosuggest
